Refactored the initial design to be more flexible.
Deleted the markov-uint file and associated interfaces. Created a State
interface that serves as keys into the Markov model. This interface is flexible
enough to represent any type of state for which this model is recording
transitions. Also added a NormalizingRow type which performs the logic necessary
for keeping a set of weights between given states and renormalizing the weights
when needed.