Transitioned the NormalizingRow to use a count instead of the weight.
Rather than try to keep a current weight, the NormalizingRow now keeps a count
of the number of times a specific state has been seen as well as the number of
total states transitioned to. This allows for counting state transitions without
having to renormalize weights while building the model.